pyproject.toml
[tool.poetry]
name = "data-quality"
version = "0.1.0"
description = ""
authors = ["Deborah Foroni <deborah.foroni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mkdocs = "^1.6.0"
mkdocs-mermaid2-plugin = "^1.1.1"
mkdocs-material = "^9.5.25"
mkdocstrings = "^0.25.1"
isort = "^5.13.2"
pytest = "^8.2.1"
tasks = "^2.8.0"
taskipy = "^1.12.2"
pydantic = "^2.7.2"
pandera = {extras = ["io"], version = "^0.19.3"}
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
format = """
isort .
black .
"""
kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
run = """
python3 app/main.py
"""
doc = "mkdocs serve"